requirements-minimal.txt
fastapi==0.121.1
uvicorn==0.38.0
websocket-client==1.9.0
Jinja2==3.0.3

